* Fix a number of gcc errors/warnings in KAI.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@2404 37977431-3df6-0310-b722-df95706aa16b

+++ AI/Global/KAI 0.11/SpotFinder.cpp
@@ -109,7 +109,7 @@ void CSpotFinder::SetRadius(int radius)
 		for (int a=0;a<DoubleRadius+1;a++){ 
 			float z=a-radius;
 			float floatsqrradius = SquareRadius;
-			xend[a]=sqrt(floatsqrradius-z*z);
+			xend[a]=int(sqrtf(floatsqrradius-z*z));
 		}
 		
 	}
